+++ src/main/java/org/ject/support/admin/auth/controller/AdminAuthController.java
@@ -1,14 +1,14 @@
-package org.ject.support.domain.admin.controller;
+package org.ject.support.admin.auth.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.ject.support.common.security.CustomSuccessHandler;
-import org.ject.support.domain.admin.dto.AdminAuthSendSlackRequest;
-import org.ject.support.domain.admin.dto.AdminAuthSendSlackResponse;
-import org.ject.support.domain.admin.dto.AdminVerifySlackRequest;
-import org.ject.support.domain.admin.service.AdminAuthService;
+import org.ject.support.admin.auth.dto.AdminAuthSendSlackRequest;
+import org.ject.support.admin.auth.dto.AdminAuthSendSlackResponse;
+import org.ject.support.admin.auth.dto.AdminVerifySlackRequest;
+import org.ject.support.admin.auth.service.AdminAuthService;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
