+++ src/test/java/org/ject/support/admin/auth/component/AdminMemberComponentTest.java
@@ -1,5 +1,6 @@
-package org.ject.support.domain.admin.component;
+package org.ject.support.admin.auth.component;
 
+import org.ject.support.admin.auth.component.AdminMemberComponent;
 import org.ject.support.base.UnitTestSupport;
 import org.ject.support.domain.admin.exception.AdminErrorCode;
 import org.ject.support.domain.admin.exception.AdminException;
